temple_checkin_app/temple_backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Werkzeug==3.0.1
python-dotenv==1.0.0
marshmallow==3.20.1
bcrypt==4.1.2
email-validator==2.1.0
python-dateutil==2.8.2
uuid==1.30
gunicorn==21.2.0